This commit was generated by cvs2svn to compensate for changes in r9133, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@9134 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/os2/archdep.c
@@ -234,7 +234,7 @@ int archdep_default_logger(const char *lvl, const char *txt)
 
 FILE *archdep_open_default_log_file()
 {
-    int val;
+    long val;
 
     char *fname = concat (archdep_boot_path(), "\\vice2.log", NULL);
     fLog = fopen(fname, "w");
